Rename action creators directory in Players module
