+++ src/Players/ActionCreators/DeletePlayer.js
